refactor: extract monitoring behavior aggregation group
