+++ src/features/monitoring/domain/monitoringDailyAnalytics.ts
@@ -27,6 +27,7 @@ import { inferGoalTagLinks, assessGoalProgress } from './goalProgressUtils';
 import type { IspRecommendationSummary } from './ispRecommendationTypes';
 import { ISP_RECOMMENDATION_LABELS } from './ispRecommendationTypes';
 import { buildIspRecommendations } from './ispRecommendationUtils';
+import { aggregateMonitoringBehaviors } from './monitoringDailyBehaviors';
 import { buildMonitoringPeriodMetrics } from './monitoringDailyPeriod';
 
 // ─── 型定義 ──────────────────────────────────────────────
@@ -133,14 +134,6 @@ const LUNCH_LABELS: Record<LunchIntake, string> = {
   none: 'なし',
 };
 
-const BEHAVIOR_LABELS: Record<ProblemBehaviorType, string> = {
-  selfHarm: '自傷',
-  otherInjury: '他傷',
-  shouting: '大声',
-  pica: '異食',
-  other: 'その他',
-};
-
 const STABLE_INTAKES: LunchIntake[] = ['full', '80'];
 
 // ─── ヘルパー ────────────────────────────────────────────
@@ -149,18 +142,6 @@ const clean = (s: unknown): string => {
   const t = String(s ?? '').trim();
   return t || '';
 };
-
-/** YYYY-MM-DD → 週番号文字列（月曜起点） */
-function toWeekKey(ymd: string): string {
-  const d = new Date(ymd + 'T00:00:00');
-  const day = d.getDay();
-  // 月曜起点に補正
-  const monday = new Date(d);
-  monday.setDate(d.getDate() - ((day + 6) % 7));
-  const mm = String(monday.getMonth() + 1).padStart(2, '0');
-  const dd = String(monday.getDate()).padStart(2, '0');
-  return `${mm}/${dd}〜`;
-}
 
 function topN(counts: Record<string, number>, n: number): ActivityRank[] {
   return Object.entries(counts)
@@ -220,85 +201,7 @@ export function aggregateLunch(records: DailyTableRecord[]): LunchSummary {
 }
 
 export function aggregateBehaviors(records: DailyTableRecord[]): BehaviorSummary {
-  const typeCounts: Record<string, number> = {};
-  const weekCounts: Record<string, number> = {};
-  let totalDays = 0;
-
-  // 問題行動があるレコードのみ集計
-  const pbRecords: DailyTableRecord[] = [];
-
-  for (const r of records) {
-    const pbs = r.problemBehaviors ?? [];
-    if (pbs.length === 0) continue;
-    totalDays++;
-    pbRecords.push(r);
-
-    for (const pb of pbs) {
-      typeCounts[pb] = (typeCounts[pb] ?? 0) + 1;
-    }
-
-    const wk = toWeekKey(r.recordDate);
-    weekCounts[wk] = (weekCounts[wk] ?? 0) + 1;
-  }
-
-  const byType = Object.entries(typeCounts)
-    .sort((a, b) => b[1] - a[1])
-    .map(([type, count]) => ({
-      type: type as ProblemBehaviorType,
-      label: BEHAVIOR_LABELS[type as ProblemBehaviorType] ?? type,
-      count,
-    }));
-
-  const weeklyTrend = Object.entries(weekCounts)
-    .sort((a, b) => a[0].localeCompare(b[0]))
-    .map(([week, count]) => ({ week, count }));
-
-  // 直近変化: 全レコード期間の中間日を基準に前半/後半で比較
-  const { recentChange, changeRate } = computeRecentChange(pbRecords, records);
-
-  const recordedDays = records.length;
-  const rate = recordedDays > 0 ? Math.round((totalDays / recordedDays) * 100) : 0;
-
-  return { totalDays, rate, byType, weeklyTrend, recentChange, changeRate };
-}
-
-/**
- * 全レコードの日付範囲の中間日を基準に前半/後半の問題行動件数を比較する。
- */
-function computeRecentChange(
-  pbRecords: DailyTableRecord[],
-  allRecords: DailyTableRecord[],
-): { recentChange: 'up' | 'down' | 'flat'; changeRate: number } {
-  if (pbRecords.length < 2) return { recentChange: 'flat', changeRate: 0 };
-  if (allRecords.length < 2) return { recentChange: 'flat', changeRate: 0 };
-
-  // 全レコード期間の中間日を算出
-  const sortedDates = allRecords.map((r) => r.recordDate).sort();
-  const from = sortedDates[0];
-  const to = sortedDates[sortedDates.length - 1];
-  const fromMs = new Date(from + 'T00:00:00').getTime();
-  const toMs = new Date(to + 'T00:00:00').getTime();
-  const midMs = fromMs + (toMs - fromMs) / 2;
-  const midDate = new Date(midMs).toISOString().slice(0, 10);
-
-  let olderCount = 0;
-  let recentCount = 0;
-  for (const r of pbRecords) {
-    if (r.recordDate <= midDate) {
-      olderCount++;
-    } else {
-      recentCount++;
-    }
-  }
-
-  if (olderCount === 0 && recentCount === 0) return { recentChange: 'flat', changeRate: 0 };
-  if (olderCount === 0) return { recentChange: 'up', changeRate: 100 };
-
-  const changeRate = Math.round(((recentCount - olderCount) / olderCount) * 100);
-
-  if (changeRate > 10) return { recentChange: 'up', changeRate };
-  if (changeRate < -10) return { recentChange: 'down', changeRate };
-  return { recentChange: 'flat', changeRate };
+  return aggregateMonitoringBehaviors(records);
 }
 
 // ─── 行動タグ集計 ────────────────────────────────────────
